+++ lib/xema.ex
@@ -12,50 +12,48 @@ defmodule Xema do
   + `__MODULE__.validate/2`
   + `__MODULE__.validate!/2`
 
-  The macro `defxema/2` supports the construction of a schema. After that
+  The macro `xema/2` supports the construction of a schema. After that
   the schema is available as a function.
 
   ## Example
 
   ```elixir
-  defmodule Schema do
-    use Xema
-
-    @pos integer(minimum: 0)
-    @neg integer(maximum: 0)
-
-    defxema user,
-            map(
-              properties: %{
-                name: string(min_length: 1),
-                age: @pos
-              }
-            )
-
-    defxema nums,
-            map(
-              properties: %{
-                pos: list(items: @pos),
-                neg: list(items: @neg)
-              }
-            )
-  end
+  iex> defmodule Schema do
+  ...>   use Xema
+  ...>
+  ...>   @pos integer(minimum: 0)
+  ...>   @neg integer(maximum: 0)
+  ...>
+  ...>   xema :user,
+  ...>        map(
+  ...>          properties: %{
+  ...>            name: string(min_length: 1),
+  ...>            age: @pos
+  ...>          }
+  ...>        )
+  ...>
+  ...>   xema :nums,
+  ...>        map(
+  ...>          properties: %{
+  ...>            pos: list(items: @pos),
+  ...>            neg: list(items: @neg)
+  ...>          }
+  ...>        )
+  ...> end
+  iex>
+  iex> Schema.valid?(:user, %{name: "John", age: 21})
+  true
+  iex> Schema.valid?(:user, %{name: "", age: 21})
+  false
+  iex> Schema.validate(:user, %{name: "John", age: 21})
+  :ok
+  iex> Schema.validate(:user, %{name: "", age: 21})
+  {:error, %{properties: %{name: %{min_length: 1, value: ""}}}}
+  iex> Schema.valid?(:nums, %{pos: [1, 2, 3]})
+  true
+  iex> Schema.valid?(:nums, %{neg: [1, 2, 3]})
+  false
   ```
-
-  The module `Schema` can now be used like this.
-
-  ```elixir
-  true = Schema.valid?(:user, %{name: "John", age: 21})
-  false = Schema.valid?(:user, %{name: "", age: 21})
-
-  :ok = Schema.validate(:user, %{name: "John", age: 21})
-  {:error, reason} = Schema.validate(:user, %{name: "", age: 21})
-
-  true = Schema.valid?(:nums, %{pos: [1, 2, 3]})
-  false = Schema.valid?(:nums, %{neg: [1, 2, 3]})
-
-  %Xema{} = Schema.user()
-  %Xema{} = Schema.nums()
   """
 
   use Xema.Behaviour
@@ -67,18 +65,11 @@ defmodule Xema do
   @keywords Schema.keywords()
   @types Schema.types()
 
+  @doc false
   defmacro __using__(_opts) do
     quote do
       import Xema.Builder
-
-      def valid?(schema, value),
-        do: Xema.valid?(apply(__MODULE__, schema, []), value)
-
-      def validate(schema, value),
-        do: Xema.validate(apply(__MODULE__, schema, []), value)
-
-      def validate!(schema, value),
-        do: Xema.validate!(apply(__MODULE__, schema, []), value)
+      @xemas []
     end
   end