+++ lib/xema/no_loader.ex
@@ -8,8 +8,9 @@ defmodule Xema.NoLoader do
   @behaviour Xema.Loader
 
   @doc """
-  Returns always the error tuple `{:error, "No loader configured."}`.
+  Returns for `any` data always the error tuple
+  `{:error, "No loader configured."}`.
   """
-  @spec fetch(any) :: {:error, binary}
+  @spec fetch(any :: any) :: {:error, binary}
   def fetch(_), do: {:error, "No loader configured."}
 end