+++ test/xema/use_test.exs
@@ -9,41 +9,29 @@ defmodule Xema.UseTest do
     @pos integer(minimum: 0)
     @neg integer(maximum: 0)
 
-    defxema user,
-            map(
-              properties: %{
-                name: string(min_length: 1),
-                age: @pos
-              }
-            )
+    xema :user,
+         map(
+           properties: %{
+             name: string(min_length: 1),
+             age: @pos
+           }
+         )
 
-    defxema person,
-            keyword(
-              properties: %{
-                name: string(min_length: 1),
-                age: @pos
-              }
-            )
+    xema :person,
+         keyword(
+           properties: %{
+             name: string(min_length: 1),
+             age: @pos
+           }
+         )
 
-    defxema nums,
-            map(
-              properties: %{
-                pos: list(items: @pos),
-                neg: list(items: @neg)
-              }
-            )
-  end
-
-  test "user/0 returns Xema" do
-    assert %Xema{} = Schema.user()
-  end
-
-  test "nums/0 returns Xema" do
-    assert %Xema{} = Schema.nums()
-  end
-
-  test "person/0 returns Xema" do
-    assert %Xema{} = Schema.person()
+    xema :nums,
+         map(
+           properties: %{
+             pos: list(items: @pos),
+             neg: list(items: @neg)
+           }
+         )
   end
 
   test "valid?/2 returns true for a valid person" do